Requestify the Read interface.
This is a big change.

Instead of having an EventHandler for Data in IUVStream, we now have
a function Read which basically creates a request and then executes it.

For every Read request there is one corresponding read in the underlying
event loop.

Currently the underlying libuv C api still uses read-stop/start, but it will
change in the future. We can emulate this future change already with
handling the Resume/Pause on our own. This simplifies the API greatly.

Becaus instead of having a Data event, Pause and Resume functions and a
ByteBufferAllocator, we have now only a single Read function, which is
easy to "Wrap" to create an Async version of it.

The future of libuv is now in LibuvSharp!